perf(server): split the projects routes and cut file write round trips
routes/projects.ts was 502 lines covering four unrelated things. It is
now routes/projects/, one file each for the project resource, files,
chat and repo generation, with requireAuth on the parent router.

Query changes, measured against us-east-2 where one round trip costs
285ms from here and about 18ms from Cloud Run:

- PATCH /:projectId/files/:fileId, 5 round trips to 4. Ownership rides
  on UPDATE ... FROM project ... RETURNING, which also yields the type
  the doc stamp needs, so the select in front of it is gone. That
  select also pulled project_file_content.spec, a TOASTed column only
  the doc branch reads, on every canvas autosave. It is now read inside
  that branch only.
- DELETE /:projectId/files/:fileId, 2 round trips to 1, via
  EXISTS on the owning project. The content row still goes with it
  through the cascade on file_id.

Verified against a throwaway Postgres with the real schema: ownership
isolation on every verb, 404 rather than an empty list for another
user's project, a real file id under the wrong project id, PATCH not
clobbering sibling content columns, explicit null still clearing one,
the doc spec stamp firing for docs and not diagrams, and the cascade.

Two known-slow routes are left alone with TODOs on them: POST files is
still 5 round trips, and PATCH could be a single CTE. Both are worth
doing but neither is on a path the user waits for.
